Improve pattern
