+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/package.scala
@@ -1,8 +1,12 @@
 package com.rasterfoundry.backsplash
 
+import cats.effect._
 import geotrellis.raster.summary._
 import io.circe.generic.semiauto._
 import io.circe.{Encoder, KeyEncoder}
+import org.http4s._
+
+import java.util.UUID
 
 package object server {
 
@@ -12,4 +16,27 @@ package object server {
   }
 
   implicit val statsEncoder: Encoder[Statistics[Double]] = deriveEncoder
+
+  val HeaderPlatNameKey = "platformName"
+  val HeaderPlatIdKey = "platformId"
+
+  implicit class ResponseIOOps(val resp: Response[IO]) extends AnyVal {
+    def addTempPlatformInfo(
+        platNameOpt: Option[String],
+        platIdOpt: Option[UUID]
+    ): Response[IO] =
+      resp
+        .putHeaders(
+          Header(
+            HeaderPlatNameKey,
+            platNameOpt.getOrElse("")
+          )
+        )
+        .putHeaders(
+          Header(
+            HeaderPlatIdKey,
+            platIdOpt.map(_.toString).getOrElse("")
+          )
+        )
+  }
 }
