+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/services/AnalysisService.scala
@@ -1,7 +1,6 @@
 package com.rasterfoundry.backsplash.server
 
 import com.rasterfoundry.backsplash.Parameters._
-import com.rasterfoundry.backsplash.utils.ResponseUtils
 import com.rasterfoundry.datamodel.UserWithPlatform
 
 import cats.effect.IO
@@ -10,7 +9,7 @@ import org.http4s.dsl.io._
 
 class AnalysisService[Param, HistStore](
     analysisManager: AnalysisManager[Param, HistStore]
-) extends ResponseUtils {
+) {
 
   val routes: AuthedRoutes[UserWithPlatform, IO] =
     AuthedRoutes.of {
@@ -18,13 +17,13 @@ class AnalysisService[Param, HistStore](
             :? NodeQueryParamMatcher(node) as user =>
         analysisManager
           .histogram(user.toUser, analysisId, node)
-          .map(addTempPlatformInfo(_, user.platformNameOpt, user.platformIdOpt))
+          .map(_.addTempPlatformInfo(user.platformNameOpt, user.platformIdOpt))
 
       case GET -> Root / UUIDWrapper(analysisId) / "statistics"
             :? NodeQueryParamMatcher(node) as user =>
         analysisManager
           .statistics(user.toUser, analysisId, node)
-          .map(addTempPlatformInfo(_, user.platformNameOpt, user.platformIdOpt))
+          .map(_.addTempPlatformInfo(user.platformNameOpt, user.platformIdOpt))
 
       case GET -> Root / UUIDWrapper(analysisId) / IntVar(z) / IntVar(
             x
@@ -32,7 +31,7 @@ class AnalysisService[Param, HistStore](
             :? NodeQueryParamMatcher(node) as user =>
         analysisManager
           .tile(user.toUser, analysisId, node, z, x, y)
-          .map(addTempPlatformInfo(_, user.platformNameOpt, user.platformIdOpt))
+          .map(_.addTempPlatformInfo(user.platformNameOpt, user.platformIdOpt))
 
       case authedReq @ GET -> Root / UUIDWrapper(analysisId) / "raw"
             :? ExtentQueryParamMatcher(extent)
@@ -47,6 +46,6 @@ class AnalysisService[Param, HistStore](
             extent,
             zoom
           )
-          .map(addTempPlatformInfo(_, user.platformNameOpt, user.platformIdOpt))
+          .map(_.addTempPlatformInfo(user.platformNameOpt, user.platformIdOpt))
     }
 }
